refactor pf statistics to new file and their own namespace
